requirements.txt
imageio==2.32.0
imutils==0.5.4
lazy_loader==0.3
networkx==3.2.1
numpy==1.26.0
opencv-python==4.8.1.78
packaging==23.2
Pillow==10.0.1
py-sudoku==1.0.3
pytesseract==0.3.10
scikit-image==0.22.0
scipy==1.11.3
tifffile==2023.9.26

easyocr~=1.7.1